feat: interfaces for PaymentService
