+++ src/core/users/DisplayPrice.ts
@@ -5,8 +5,3 @@ export interface DisplayPrice {
   amount: number;
   currency: string;
 }
-
-export interface RequestedPlan {
-  selectedPlan: DisplayPrice & { decimalAmount: number };
-  upsellPlan?: DisplayPrice & { decimalAmount: number };
-}
